Move OtelTraceLayer into otel_util
